requirements.txt
streamlit
clickhouse-driver
plotly
pandas
python-dotenv